Added front-end POST request, fixed typo in Express application.
Added form action to index.html that sends a POST request to the back-end when a form is submitted on the webpage. Added a console log for the Express application that prints a confirmation when data is received.
Also fixed a small typo in app.js. "bodyparser" was used instead of "bodyParser"